requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# AST Parsing and Analysis
tree-sitter>=0.20.0
radon>=6.0.1
astroid>=2.15.0

# Deep Learning
torch>=2.0.0
transformers>=4.30.0
tensorflow>=2.13.0

# Clustering and Dimensionality Reduction
umap-learn>=0.5.3
hdbscan>=0.8.33

# Code Analysis
lizard>=1.17.10
pylint>=2.17.0

# Data Handling
requests>=2.31.0
tqdm>=4.65.0
jsonlines>=3.1.0

# Visualization
plotly>=5.14.0
networkx>=3.1

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
